[Fix] fix database type define error

+++ src/server/kuwo/web_api/lyric.rs
@@ -20,7 +20,7 @@ struct GetLrc {
 }
 
 fn format_seconds_to_timestamp(seconds: f64) -> String {
-    let minutes = (seconds / 60.0).floor() as i32;
+    let minutes = (seconds / 60.0).floor() as i64;
     let remaining_seconds = seconds % 60.0;
     format!("{:02}:{:05.2}", minutes, remaining_seconds)
 }
